+++ src/components/Chat/Chat.tsx
@@ -3,9 +3,10 @@ import * as PropTypes from 'prop-types'
 import * as React from 'react'
 
 import { childrenExist, customPropTypes, UIComponent } from '../../lib'
+import ChatItem from './ChatItem'
 import ChatMessage from './ChatMessage'
-import { ComponentVariablesInput, ComponentPartStyle } from '../../../types/theme'
-import { Extendable, ReactChildren, ItemShorthand } from '../../../types/utils'
+import { ComponentPartStyle, ComponentVariablesInput } from '../../../types/theme'
+import { Extendable, ItemShorthand, ReactChildren } from '../../../types/utils'
 import { Accessibility, AccessibilityActionHandlers } from '../../lib/accessibility/interfaces'
 import ChatBehavior from '../../lib/accessibility/Behaviors/Chat/ChatBehavior'
 
@@ -14,7 +15,7 @@ export interface IChatProps {
   as?: any
   className?: string
   children?: ReactChildren
-  messages?: ItemShorthand[]
+  items?: ItemShorthand[]
   styles?: ComponentPartStyle
   variables?: ComponentVariablesInput
 }
@@ -28,15 +29,20 @@ class Chat extends UIComponent<Extendable<IChatProps>, any> {
     /** Accessibility behavior if overridden by the user. */
     accessibility: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
 
+    /** An element type to render as (string or function). */
     as: customPropTypes.as,
 
     /** Additional CSS class name(s) to apply.  */
     className: PropTypes.string,
 
+    /**
+     *  Used to set content when using childrenApi - internal only
+     *  @docSiteIgnore
+     */
     children: PropTypes.node,
 
-    /** Shorthand array of messages. */
-    messages: PropTypes.arrayOf(PropTypes.any),
+    /** Shorthand array of the items inside the chat. */
+    items: PropTypes.arrayOf(customPropTypes.itemShorthand),
 
     /** Additional CSS styles to apply to the component instance.  */
     styles: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
@@ -45,18 +51,9 @@ class Chat extends UIComponent<Extendable<IChatProps>, any> {
     variables: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
   }
 
-  static handledProps = [
-    'accessibility',
-    'as',
-    'children',
-    'className',
-    'messages',
-    'styles',
-    'variables',
-  ]
-
   static defaultProps = { accessibility: ChatBehavior as Accessibility, as: 'ul' }
 
+  static Item = ChatItem
   static Message = ChatMessage
 
   actionHandlers: AccessibilityActionHandlers = {
@@ -64,7 +61,7 @@ class Chat extends UIComponent<Extendable<IChatProps>, any> {
   }
 
   renderComponent({ ElementType, classes, accessibility, rest }) {
-    const { children, messages } = this.props
+    const { children, items } = this.props
 
     return (
       <ElementType
@@ -73,9 +70,7 @@ class Chat extends UIComponent<Extendable<IChatProps>, any> {
         {...accessibility.keyHandlers.root}
         {...rest}
       >
-        {childrenExist(children)
-          ? children
-          : _.map(messages, message => ChatMessage.create(message))}
+        {childrenExist(children) ? children : _.map(items, item => ChatItem.create(item))}
       </ElementType>
     )
   }