+++ src/components/List/List.tsx
@@ -32,6 +32,10 @@ class List extends UIComponent<Extendable<IListProps>, any> {
   static propTypes = {
     as: customPropTypes.as,
 
+    /**
+     *  Used to set content when using childrenApi - internal only
+     *  @docSiteIgnore
+     */
     children: PropTypes.node,
 
     /** Additional CSS class name(s) to apply.  */
@@ -66,20 +70,6 @@ class List extends UIComponent<Extendable<IListProps>, any> {
     as: 'ul',
     accessibility: ListBehavior as Accessibility,
   }
-
-  static handledProps = [
-    'accessibility',
-    'as',
-    'children',
-    'className',
-    'debug',
-    'items',
-    'selection',
-    'styles',
-    'truncateContent',
-    'truncateHeader',
-    'variables',
-  ]
 
   static Item = ListItem