+++ src/components/Menu/Menu.tsx
@@ -47,7 +47,10 @@ class Menu extends AutoControlledComponent<Extendable<IMenuProps>, any> {
     /** Index of the currently active item. */
     activeIndex: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
 
-    /** Primary content. */
+    /**
+     *  Used to set content when using childrenApi - internal only
+     *  @docSiteIgnore
+     */
     children: PropTypes.node,
 
     /** Additional CSS class name(s) to apply.  */
@@ -97,25 +100,6 @@ class Menu extends AutoControlledComponent<Extendable<IMenuProps>, any> {
     as: 'ul',
     accessibility: MenuBehavior as Accessibility,
   }
-
-  static handledProps = [
-    'accessibility',
-    'activeIndex',
-    'as',
-    'children',
-    'className',
-    'defaultActiveIndex',
-    'fluid',
-    'iconOnly',
-    'items',
-    'pills',
-    'pointing',
-    'styles',
-    'type',
-    'underlined',
-    'variables',
-    'vertical',
-  ]
 
   static autoControlledProps = ['activeIndex']