+++ src/components/Menu/MenuItem.tsx
@@ -7,6 +7,7 @@ import { childrenExist, createShorthandFactory, customPropTypes, UIComponent } f
 import Icon from '../Icon'
 import { MenuItemBehavior } from '../../lib/accessibility'
 import { Accessibility, AccessibilityActionHandlers } from '../../lib/accessibility/interfaces'
+import IsFromKeyboard from '../../lib/isFromKeyboard'
 
 import { ComponentVariablesInput, ComponentPartStyle } from '../../../types/theme'
 import {
@@ -37,7 +38,11 @@ export interface IMenuItemProps {
   variables?: ComponentVariablesInput
 }
 
-class MenuItem extends UIComponent<Extendable<IMenuItemProps>, any> {
+export interface IMenuItemState {
+  [IsFromKeyboard.propertyName]: boolean
+}
+
+class MenuItem extends UIComponent<Extendable<IMenuItemProps>, IMenuItemState> {
   static displayName = 'MenuItem'
 
   static className = 'ui-menu__item'
@@ -51,7 +56,10 @@ class MenuItem extends UIComponent<Extendable<IMenuItemProps>, any> {
     /** An element type to render as (string or function). */
     as: customPropTypes.as,
 
-    /** Primary content. */
+    /**
+     *  Used to set content when using childrenApi - internal only
+     *  @docSiteIgnore
+     */
     children: PropTypes.node,
 
     /** Additional CSS class name(s) to apply.  */
@@ -114,34 +122,7 @@ class MenuItem extends UIComponent<Extendable<IMenuItemProps>, any> {
     accessibility: MenuItemBehavior as Accessibility,
   }
 
-  static handledProps = [
-    'accessibility',
-    'active',
-    'as',
-    'children',
-    'className',
-    'content',
-    'disabled',
-    'icon',
-    'iconOnly',
-    'index',
-    'onClick',
-    'pills',
-    'pointing',
-    'styles',
-    'type',
-    'underlined',
-    'variables',
-    'vertical',
-  ]
-
-  protected actionHandlers: AccessibilityActionHandlers = {
-    performClick: event => this.handleClick(event),
-  }
-
-  handleClick = e => {
-    _.invoke(this.props, 'onClick', e, this.props)
-  }
+  state = IsFromKeyboard.initial
 
   renderComponent({ ElementType, classes, accessibility, rest }) {
     const { children, content, icon } = this.props
@@ -159,6 +140,8 @@ class MenuItem extends UIComponent<Extendable<IMenuItemProps>, any> {
           <a
             className={cx('ui-menu__item__anchor', classes.anchor)}
             onClick={this.handleClick}
+            onBlur={this.handleBlur}
+            onFocus={this.handleFocus}
             {...accessibility.attributes.anchor}
             {...accessibility.keyHandlers.anchor}
           >
@@ -171,6 +154,26 @@ class MenuItem extends UIComponent<Extendable<IMenuItemProps>, any> {
         )}
       </ElementType>
     )
+  }
+
+  protected actionHandlers: AccessibilityActionHandlers = {
+    performClick: event => this.handleClick(event),
+  }
+
+  private handleClick = e => {
+    _.invoke(this.props, 'onClick', e, this.props)
+  }
+
+  private handleBlur = (e: React.SyntheticEvent) => {
+    this.setState(IsFromKeyboard.initial)
+
+    _.invoke(this.props, 'onBlur', e, this.props)
+  }
+
+  private handleFocus = (e: React.SyntheticEvent) => {
+    this.setState(IsFromKeyboard.state())
+
+    _.invoke(this.props, 'onFocus', e, this.props)
   }
 }