+++ test/specs/components/Chat/ChatItem-test.tsx
@@ -1,0 +1,7 @@
+import { isConformant } from 'test/specs/commonTests'
+
+import ChatItem from 'src/components/Chat/ChatItem'
+
+describe('ChatItem', () => {
+  isConformant(ChatItem)
+})